storage/SDK separation phase I
