+++ aim/storage/encoding/__init__.py
@@ -1,0 +1,1 @@
+from aim.storage.encoding.encoding import *
