+++ aim/storage/hashing/__init__.py
@@ -1,0 +1,1 @@
+from aim.storage.hashing.hashing import hash_auto
