+++ aim/storage/hashing/c_hash.pxd
