+++ aim/storage/hashing/c_hash.pyx
@@ -1,6 +1,6 @@
 # distutils: language = c++
 
-from aim.storage cimport c_hash
+from aim.storage.hashing cimport c_hash
 
 cpdef Py_hash_t det_hash(v):
     return c_hash.le_Py_HashDouble(v, v)
