+++ aim/storage/sdk/__init__.py
@@ -1,0 +1,1 @@
+# TODO: move package to top-level
