+++ aim/storage/sdk/run.py
@@ -5,19 +5,20 @@ from time import time
 from collections import Counter
 
 from aim.storage.types import AimObject
-from aim.storage.trace import RunTraceCollection
+from aim.storage.sdk.trace import RunTraceCollection
 from aim.storage.hashing import hash_auto
 from aim.storage.context import Context, Metric
 from aim.storage.treeview import TreeView
 from aim.storage.containerview import ContainerView
+from aim.storage.proxy import AimObjectProxy
 
 from typing import Any, Dict, Iterator, Optional, Tuple
 from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
-    from aim.storage.trace import Trace
-    from aim.storage.trace import TraceCollection
-    from aim.storage.repo import Repo
+    from aim.storage.sdk.trace import Trace
+    from aim.storage.sdk.trace import TraceCollection
+    from aim.storage.sdk.repo import Repo
 
 
 logger = logging.getLogger(__name__)
@@ -30,7 +31,7 @@ class Run:
 
     def __init__(self, hashname: str, *, repo: "Repo" = None, read_only: bool = False):
         if repo is None:
-            from aim.storage.repo import Repo
+            from aim.storage.sdk.repo import Repo
             repo = Repo.default_repo()
 
         self.repo = repo
@@ -127,6 +128,9 @@ class Run:
         epoch_view[step] = epoch
         time_view[step] = time()
 
+    def proxy(self):
+        return AimObjectProxy(lambda: self.meta_run_tree, view=self.meta_run_tree)
+
     def trace_tree(self, name: str, context: Context) -> TreeView:
         return self.series_run_tree.view((context.idx, name))
 
@@ -151,7 +155,7 @@ class Run:
             metric_name: str,
             context: Context
     ) -> Optional['Trace']:
-        from aim.storage.trace import Trace
+        from aim.storage.sdk.trace import Trace
         trace = Trace(metric_name, context, self)
         return trace if bool(trace) else None
